+++ lib/index.ts
@@ -10,37 +10,43 @@ import identityContext from "./authorizer/model/identityContext";
 import policyContext from "./authorizer/model/policyContext";
 import policyInstance from "./authorizer/model/policyInstance";
 import queryOptions from "./authorizer/model/queryOptions";
+import {
+  Directory,
+  DirectoryConfig,
+  DirectoryServiceV2,
+  ds,
+} from "./directory/v2";
 import { DirectoryServiceV3, DirectoryV3 } from "./directory/v3";
 import { DirectoryV3Config } from "./directory/v3/types";
 import { displayStateMap } from "./displayStateMap";
-import { Directory, DirectoryConfig, ds } from "./ds";
 import { is } from "./is";
 import { AuthzOptions, jwtAuthz } from "./jwtAuthz";
 import getSSLCredentials from "./ssl";
 export {
-  is,
-  jwtAuthz,
-  displayStateMap,
-  ds,
-  Directory,
-  Authorizer,
-  Middleware,
-  authz,
-  SubIdentityMapper,
-  JWTIdentityMapper,
   AnonymousIdentityMapper,
-  ManualIdentityMapper,
-  DirectoryConfig,
+  Authorizer,
+  authz,
   AuthzOptions,
-  PolicyPathMapper,
-  ObjectIDFromVar,
-  identityContext,
-  policyContext,
-  policyInstance,
-  getSSLCredentials,
-  queryOptions,
   decisionTreeOptions,
+  Directory,
+  DirectoryConfig,
+  DirectoryServiceV2,
   DirectoryServiceV3,
   DirectoryV3,
   DirectoryV3Config,
+  displayStateMap,
+  ds,
+  getSSLCredentials,
+  identityContext,
+  is,
+  jwtAuthz,
+  JWTIdentityMapper,
+  ManualIdentityMapper,
+  Middleware,
+  ObjectIDFromVar,
+  policyContext,
+  policyInstance,
+  PolicyPathMapper,
+  queryOptions,
+  SubIdentityMapper,
 };
